refactor: improve os command
